+++ src/Revit.TestRunner.Shared/Communication/Server/Request.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.IO;
-using Revit.TestRunner.Shared.Communication.Dto;
+using Revit.TestRunner.Shared.Dto;
 
 namespace Revit.TestRunner.Shared.Communication.Server
 {