+++ src/Revit.TestRunner.Shared/Dto/NoParameterDto.cs
@@ -1,0 +1,12 @@
+﻿namespace Revit.TestRunner.Shared.Dto
+{
+    /// <summary>
+    /// Simple Get call Dto. No Parameters.
+    /// </summary>
+    public class NoParameterDto : BaseRequestDto
+    {
+        public NoParameterDto() : base( DtoType.NoParameterDto )
+        {
+        }
+    }
+}