+++ src/Revit.TestRunner.Shared/Dto/TestCaseDto.cs
@@ -1,0 +1,24 @@
+﻿using Revit.TestRunner.Shared.Communication;
+
+namespace Revit.TestRunner.Shared.Dto
+{
+    /// <summary>
+    /// Represents a TestCase / Test method.
+    /// </summary>
+    public class TestCaseDto
+    {
+        public string Id { get; set; }
+
+        public string AssemblyPath { get; set; }
+
+        public string TestClass { get; set; }
+
+        public string MethodName { get; set; }
+
+        public TestState State { get; set; }
+
+        public string Message { get; set; }
+
+        public string StackTrace { get; set; }
+    }
+}