+++ src/Revit.TestRunner.Shared/Dto/TestResponseDto.cs
@@ -1,0 +1,24 @@
+﻿using Newtonsoft.Json;
+
+namespace Revit.TestRunner.Shared.Dto
+{
+    /// <summary>
+    /// Response dto for a test run request.
+    /// Conains information about the run.
+    /// </summary>
+    public class TestResponseDto : BaseResponseDto
+    {
+        public TestResponseDto() : base( DtoType.TestResponseDto )
+        {
+        }
+
+        [JsonProperty( Order = 11 )]
+        public string ResponseDirectory { get; set; }
+
+        [JsonProperty( Order = 12 )]
+        public string ResultFile { get; set; }
+
+        [JsonProperty( Order = 13 )]
+        public string SummaryFile { get; set; }
+    }
+}